backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
faiss-cpu
langchain
